moved stuff around and made simple setup.py to create package structure
