+++ src/test_env.py
@@ -3,6 +3,7 @@ import numpy as np
 
 from envs import BoardEnv
 
+
 # list of source positions
 sources = np.array([[0,0]])
